requirements.txt
attrs>=17.2.0
polling>=0.3.0
requests>=2.18.1
setuptools>=36.0.1
attr>=0.2.0
requests_mock>=1.3.0
PyYAML>=3.12

